backend/requirements.txt
aiofiles
aiosqlite
aasyncpg
abcrypt
abeautifulsoup4
acython
adotenv
afastapi[standard]
afastembed
alxml
amobi
apasslib[bcrypt]
apathlib
apsycopg2-binary
apydantic
apython-jose[cryptography]
apython-multipart
arequests
asqlalchemy
auvicorn
